muda banco para sqlserver, finaliza service balance

+++ src/Services/Application/FluxoCaixa.Application/Objects/Requests/PaymentRequest.cs
@@ -8,5 +8,5 @@ public class PaymentRequest
     public string Description { get; set; }
     public DateTime PaymentDate { get; set; }
     public int PaymentType { get; set; }
-    public double PaymentValue { get; set; }
+    public decimal PaymentValue { get; set; }
 }
